refactor: 패키지 구조를 일관성 있게 수정
패키지 구조를 일관성있게 수정하였습니다.
