+++ src/main/java/chess/dao/DBConnector.java
@@ -1,4 +1,4 @@
-package chess.web.dao;
+package chess.dao;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
