+++ src/test/java/chess/web/dao/ChessGameDaoTest.java
@@ -2,10 +2,13 @@ package chess.web.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.dao.ChessGameDao;
+import chess.dao.DBConnector;
+import chess.dao.PieceDao;
 import chess.domain.ChessGame;
 import chess.domain.Command;
 import chess.domain.state.State;
-import chess.web.dto.ChessGameDto;
+import chess.dto.ChessGameDto;
 import java.sql.Connection;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
