+++ src/test/java/chess/web/dao/PieceDaoTest.java
@@ -1,7 +1,9 @@
 package chess.web.dao;
 
+import chess.dao.ChessGameDao;
+import chess.dao.PieceDao;
 import chess.domain.ChessGame;
-import chess.web.dto.ChessGameDto;
+import chess.dto.ChessGameDto;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
